Sonar-Klasse hinzugefügt. sonarTest() zum Test beigelegt.

+++ src/horst/Bewegungen.java
@@ -49,8 +49,8 @@ public class Bewegungen {
 	 * Drehung durch gegensätzliches Rotieren der Achsen.
 	 * @param degree Gradanzahl die gedreht werden soll, positiv für Rechts, negativ für Links
 	 */
-	public void turn(short degree) {
-		short turn_ammount = (short) (360 / degree);
+	public void turn(int degree) {
+		int turn_ammount = 360 / degree;
 		// Geht von dem Fall aus, dass A sich Links von Fahrtrichtung befindet und B rechts davon.
 		Motor.A.rotate(-((KETTEN_UMFANG / turn_ammount) / RAD_UMFANG) * 360);
 		Motor.B.rotate(((KETTEN_UMFANG / turn_ammount) / RAD_UMFANG) * 360);
